spec: harmonize FIELD_BINARY fields
esp. the 310 fields.
FIELD_BINARY falls back to FIELD_TF elsewhere,
and is needed for dxf *and* json. json cannot print \0.

+++ src/out_dxf.c
@@ -102,7 +102,7 @@ static void dxf_fixup_string (Bit_Chain *restrict dat, char *restrict str);
   }
 #define VALUE_BINARY(value, size, dxf)                                        \
   {                                                                           \
-    long len = size;                                                          \
+    long len = (long)size;                                                    \
     do                                                                        \
       {                                                                       \
         short j;                                                              \
